requirements.txt
Flask
bscscan-python
pythonpancakes
httpx
aiohttp
asyncio